Fix MSVC compiler warnings
reg.: 'return': conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/dom/chordline.cpp
@@ -275,7 +275,7 @@ int ChordLine::subtype() const
     size_t h2 = std::hash<bool> {}(m_straight);
     size_t h3 = std::hash<bool> {}(m_wavy);
 
-    return h1 ^ (h2 << 1) ^ (h3 << 2);
+    return static_cast<int>(h1 ^ (h2 << 1) ^ (h3 << 2));
 }
 
 muse::TranslatableString ChordLine::subtypeUserName() const
